requirements.txt
streamlit>=1.30.0
google-genai>=0.1.0
pandas
plotly
-e .
